Remove unused folder 'playwright-report', 'test-results', 'tests-examples'
This commit removes the 'playwright-report', 'test-results', 'tests-examples' folder, which was no longer being used in the project. It contained outdated resources that are no longer relevant to the current state of the project.

Reason for deletion:
- The assets in this folder were not used due to not using playwright anymore for testing
- Cleaning up unnecessary files to maintain a tidy project structure.

This change does not affect any functionality but helps declutter the project directory and improves maintainability.